changed AlaUserProfile to an interface and AlaOidcUserProfile now extends OidcProfile
